packages/odoo12-addon-somconnexio/odoo12_addon_somconnexio-12.0.0.0.0rc57-py2.py3-none-any.whl/odoo/addons/somconnexio/otrs_factories/customer_data_from_res_partner.py
from otrs_somconnexio.otrs_models.customer_data import CustomerData


class CustomerDataFromResPartner:

    def __init__(self, partner):
        self.partner = partner

    def build(self):
        return CustomerData(
            id=self.partner.id,
            vat_number=self.partner.vat,
            email=self.partner.email,
            phone=self.partner.mobile or self.partner.phone,
            name=self.partner.name,
            first_name=self.partner.name,
            street=self.partner.full_street,
            zip=self.partner.zip,
            city=self.partner.city,
            subdivision="{}-{}".format(
                self.partner.country_id.code,
                self.partner.state_id.code
            ),
        )
